+++ examples/demos/point/fill.ts
@@ -43,10 +43,14 @@ export const fill: TestCase = async (options) => {
         field: 'unit_price',
         value: () => 1,
       },
+
       strokeWidth: 2,
     });
 
   scene.addLayer(pointLayer);
+  // scene.on('mousemove', console.log);
+  scene.on('mouseout', console.log);
+  scene.on('mouseover', console.log);
 
   return scene;
 };